% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakdown_budget.R
\name{breakdown_budget}
\alias{breakdown_budget}
\title{Break down the budget}
\usage{
breakdown_budget(data, group_var)
}
\arguments{
\item{data}{A data set}

\item{group_var}{A grouping variable}
}
\value{
A total amount of budget expenditure by grouping variable
}
\description{
Break down the budget
}
