% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakdown_budget.R
\name{visualize_breakdown}
\alias{visualize_breakdown}
\title{Visualize the budget breakdown}
\usage{
visualize_breakdown(data, group_var)
}
\arguments{
\item{data}{A data set}

\item{group_var}{A grouping variable}
}
\value{
A ggplot
}
\description{
Visualize the budget breakdown
}
